OK FOR NOW

+++ MyEbuyWebParent/MyEbuyBackEnd/src/main/java/org/nir/myebuy/admin/user/RoleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends CrudRepository<Role, Long> 
+public interface RoleRepository extends CrudRepository<Role, Integer> 
 {
